Add account management
All admins can delete employees accounts. This does not affect the
Employees table because deleting an employee causes inconsistencies in the
database, what it does is remove the user from the users database. This
way the employee can no longer access the site. Also modified the
Employee/Index view to replace the link for a button.